# docs/phase1/phase1-openapi.yaml
openapi: 3.0.3
info:
  title: Subject Media Partner Platform API
  version: 1.0.0
  description: Phase 1 API contract for Agency Hub and Client Partner Portal.
servers:
  - url: https://api.subjectmedia.com/v1
  - url: http://localhost:3000/api
...
paths:
...
  /billing/create-invoice:
    post:
      tags: [Billing]
      summary: Create Stripe invoice for onboarding or recurring service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '200':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
...
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns in current tenant scope
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: clientOrgId
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Campaign'
    post:
      tags: [Campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCampaignRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
...
  /track/event:
    post:
      tags: [Tracking]
      summary: Public tracking ingestion endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackEventRequest'
      responses:
        '200':
          description: Tracked
...
  /leads/{leadId}/convert:
    post:
      tags: [Leads]
      summary: Mark lead as converted and attach revenue
      parameters:
        - $ref: '#/components/parameters/LeadId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revenueAmount]
              properties:
                revenueAmount:
                  type: number
                  minimum: 0
      responses:
        '200':
          description: Converted
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  parameters:
...
    LeadId:
      in: path
      name: leadId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required: [clientOrgId, amount, currency]
      properties:
        clientOrgId:
          type: string
          format: uuid
        amount:
          type: number
          minimum: 0
        currency:
          type: string
    InvoiceResponse:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
        stripeInvoiceId:
          type: string
        status:
          type: string
          enum: [draft, open, paid, void, uncollectible]
...
    Campaign:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientOrgId:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, archived]
        budgetAmount:
          type: number
          minimum: 0
    CreateCampaignRequest:
      type: object
      required: [clientOrgId, name]
      description: If both startsOn and endsOn are provided, startsOn must be less than or equal to endsOn.
      properties:
        clientOrgId:
          type: string
          format: uuid
        name:
          type: string
        goalSummary:
          type: string
        startsOn:
          type: string
          format: date
        endsOn:
          type: string
          format: date
...
    TrackEventRequest:
      type: object
      required: [type, campaignId]
      description: assetId, when provided, must belong to the same client organization as campaignId.
      properties:
        type:
          type: string
          enum: [click, scan, pageview, lead_submit, conversion]
        campaignId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
        source:
          type: string
        referrer:
          type: string
        deviceType:
          type: string
        meta:
          type: object
...
security:
  - SessionAuth: []